// frontend/src/api/generated/models/PerformanceTrend.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Performance over time.
 */
export type PerformanceTrend = {
    date: string;
    reviews_count: number;
    accuracy: number;
    study_time_minutes: number;
};